refactor using cogs
