+++ requirements.txt
@@ -1,4 +1,5 @@
-oauth2client==4.1.3
 discord.py==1.3.3
+oauth2client==4.1.3
 gspread==3.6.0
+cogs==0.4.4
 discord==1.0.1
